pyproject.toml
[project]
name = "mqtt-sensehat"
version = "0.1.0"
description = ""
authors = [
    {name = "rippleFCL",email = "github@ripple.contact"}
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "paho-mqtt>=2.1.0,<3.0.0",
    "sense-hat>=2.6.0,<3.0.0",
    "pyyaml>=6.0.2,<7.0.0",
    "pydantic>=2.11.7,<3.0.0",
    "rtimulib @ git+https://github.com/RPi-Distro/RTIMULib.git#subdirectory=Linux/python"
]

[tool.poetry]
package-mode = false

[tool.poetry.group.dev.dependencies]
ruff = "^0.12.4"
pre-commit = "^4.2.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "W1", "W2", "F", "RUF", "I"]

[tool.pyright]
venvPath = "."
venv = ".venv"
